Update andersen.odd
I debugged the facsimile error in the little Mermaid. 1) CAUSE This is an input flaw on my side with only a single set of tiff-files. 2) FAILED SOLUTION: I tried to rescale the images but did not succeed in eliminating the error 3) SOLUTION The 403-error was solved by tinkering with the ODD. I reduced the size of the request sent to the Cantaloupe image server from 1200 to 1050 px width. Cf. GitHub for details.